requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.29
asyncpg==0.29.0
psycopg[binary]==3.1.19
alembic==1.13.1
pydantic==2.6.4
boto3==1.34.67
python-multipart==0.0.9
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
celery==5.3.6
redis==5.0.1
Pillow==10.2.0
python-dotenv==1.0.1

